// src/data/search-controls.enum.ts
export enum SearchControls {
    PLACEHOLDER_ES = 'Buscar en Wikipedia',
    SEARCH_BTN_ES = 'Buscar',

    PLACEHOLDER_PT = 'Procurar em Wikipedia',
    SEARCH_BTN_PT = 'Procurar',

    PLACEHOLDER_FR = 'Rechercher sur Wikipedia',
    SEARCH_BTN_FR = 'Rechercher',

    PLACEHOLDER_DE = 'Wikipedia durchsuchen',
    SEARCH_BTN_DE = 'Suchen',

    PLACEHOLDER_UK = 'Пошук у Wikipedia',
    SEARCH_BTN_UK = 'Знайти',

    PLACEHOLDER_PL = 'Przeszukaj Wikipedię',
    SEARCH_BTN_PL = 'Szukaj',
}